Revert "Arreglo labels correspondientes e implemento rutas con 'arroba'"
