--- src/form/InputFieldTextKUI/InputFieldKUI.stories.tsx
@@ -1,12 +1,0 @@
-import { InputFieldTextKUI } from '@KUI-form';
-
-export default {
-  title: 'Form/InputFieldKUI',
-  component: InputFieldTextKUI,
-};
-
-export const Example = {
-  args: {
-    label: 'Nombre',
-  },
-};
