+++ src/form/base/index.ts
@@ -1,4 +1,1 @@
 export * from './CurrencyInputKUI';
-export * from './CheckKUI';
-export * from './PasswordInputKUI';
-export * from './LabelKUI';
