+++ src/molecules/InputFieldKUI/InputFieldKUI.stories.tsx
@@ -1,0 +1,13 @@
+import { InputFieldKUI } from '../InputFieldKUI';
+
+export default {
+  component: InputFieldKUI,
+};
+
+let estado;
+export const Example = {
+  args: {
+    label: 'Nombre',
+    value: estado,
+  },
+};
